The CSS Styles
Perhaps overkill with styling each particular line in the table but it is a beginner project where the expectations were to cram in as much as you can show so its it a nuclear option: sure. There are much tidier solutions with either Bootstrap if you want something out of the box or even basic css flexbox, nevertheless, trying to show I understood the assignment.